requirements.txt
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-bio==1.0.2
pandas==1.5.3
psycopg2-binary==2.9.5
SQLAlchemy==2.0.4
rdkit==2022.9.5
python-dotenv==1.0.0